back/requirements.txt
fastapi==0.104.1
pydantic==2.4.2
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.39
asyncpg==0.30.0
passlib==1.7.4
python-jose==3.4.0
redis==5.2.1
celery==5.4.0
google-generativeai==0.3.1